requirements/_base.txt
Django==1.8.5
django-extensions==1.5.7
djangorestframework==3.2.4
drf-ujson==1.2
ipython==4.0.0
PyJWT==1.4.0
requests==2.7.0
PyMarvel==0.1.0
django-rest-swagger==0.3.4
drf-extensions==0.2.8